Reorganize options
