+++ src/ui/options/components/edit-options/edited-tracks.tsx
