+++ src/ui/options/components/options/options.tsx
@@ -1,12 +1,9 @@
 import * as BrowserStorage from '@/core/storage/browser-storage';
 import { t } from '@/util/i18n';
-import { Setter, createResource } from 'solid-js';
+import { createResource } from 'solid-js';
 import GlobalOptionsList from './global-options';
 import ConnectorOptionsList from './connector-options';
 import ScrobbleBehavior from './scrobble-behavior';
-import EditedTracks from './edited-tracks';
-import ConnectorOverrideOptions from './connector-override';
-import RegexEdits from './regex-edits';
 
 const globalOptions = BrowserStorage.getStorage(BrowserStorage.OPTIONS);
 
@@ -17,25 +14,13 @@ const [options, setOptions] = createResource(
 /**
  * Component that shows options to the user
  */
-export default function OptionsComponent(props: {
-	setActiveModal: Setter<string>;
-	modal: HTMLDialogElement | undefined;
-}) {
+export default function OptionsComponent() {
 	return (
 		<>
 			<h1>{t('optionsOptions')}</h1>
 			<GlobalOptionsList options={options} setOptions={setOptions} />
 			<ConnectorOptionsList />
 			<ScrobbleBehavior options={options} setOptions={setOptions} />
-			<EditedTracks
-				setActiveModal={props.setActiveModal}
-				modal={props.modal}
-			/>
-			<RegexEdits
-				setActiveModal={props.setActiveModal}
-				modal={props.modal}
-			/>
-			<ConnectorOverrideOptions />
 		</>
 	);
 }
